+++ src/pages/mypage/EcoDealReservation.jsx
@@ -1,6 +1,6 @@
 import { useEffect, useMemo, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { fetchMyEcoDeals } from "@/api/order/order.api";
+import { fetchMyEcoDeals } from "@/api/member/member.api";
 
 export default function EcoDealReservation() {
   const [data, setData] = useState([]);
